refactor(pwa): don't show mail details as popover
