+++ frontend/components.d.ts
@@ -45,6 +45,7 @@ declare module 'vue' {
     MailActions: typeof import('./src/components/MailActions.vue')['default']
     MailDataExchangeSettings: typeof import('./src/components/Settings/MailDataExchangeSettings.vue')['default']
     MailDate: typeof import('./src/components/MailDate.vue')['default']
+    MailDetails: typeof import('./src/components/MailDetails.vue')['default']
     MailDetailsPopover: typeof import('./src/components/MailDetailsPopover.vue')['default']
     MailListItem: typeof import('./src/components/MailListItem.vue')['default']
     MailLogo: typeof import('./src/components/Icons/MailLogo.vue')['default']
